backend/visual_noun/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.18.0
google-adk>=1.2.1
google-genai>=1.14.0
python-dotenv>=1.0.0
